minor refactoring of views
