+++ web/core/components/pages/list/applied-filters/index.ts
@@ -1,3 +1,1 @@
-export * from "./date";
-export * from "./members";
 export * from "./root";
